Errors if grid_size bigger in global array at index (num_elements / grid_size)
I think I do not fully understand grid striding yet

```
$ nvcc -gencode=arch=compute_86,code=sm_86 compressed.cu -o compressed && ./compressed 2048 1024 1
Result of GPU is the same as CPU result.
Result of GPU is the same as CPU result.
Result of GPU is the same as CPU result.
num_elements: 2048, block size: 1024, grid_size: 1, runtime: 14.802261ms

$ nvcc -gencode=arch=compute_86,code=sm_86 compressed.cu -o compressed && ./compressed 4096 1024 1
Result of GPU is the same as CPU result.
Result of GPU is the same as CPU result.
Result of GPU is the same as CPU result.
num_elements: 4096, block size: 1024, grid_size: 1, runtime: 29.537621ms

$ nvcc -gencode=arch=compute_86,code=sm_86 compressed.cu -o compressed && ./compressed 8192 1024 1
Result of GPU is the same as CPU result.
Result of GPU is the same as CPU result.
Result of GPU is the same as CPU result.
num_elements: 8192, block size: 1024, grid_size: 1, runtime: 59.175018ms

$ nvcc -gencode=arch=compute_86,code=sm_86 compressed.cu -o compressed && ./compressed 8192 1024 2
Result 4097 of GPU is not the same as CPU result:
6711255218071955582 + 4772134560539550557 = 11483389778611506139 (host)
6711255218071955582 + 4772134560539550557 = 0 (device)
10011111 01011101 00110010 00101001 01111011 01111101 10001011 11011011
00000000 00000000 00000000 00000000 00000000 00000000 00000000 00000000

$ nvcc -gencode=arch=compute_86,code=sm_86 compressed.cu -o compressed && ./compressed 512 1024 4
Result 129 of GPU is not the same as CPU result:
5559770857505913902 + 1678463285158154560 = 7238234142664068462 (host)
5559770857505913902 + 1678463285158154560 = 0 (device)
01100100 01110011 01011111 10011010 01011010 10011000 10001101 01101110
00000000 00000000 00000000 00000000 00000000 00000000 00000000 00000000

$ nvcc -gencode=arch=compute_86,code=sm_86 compressed.cu -o compressed && ./compressed 512 1024 2
Result 257 of GPU is not the same as CPU result:
2751555560243478560 + 5939149587220868940 = 8690705147464347500 (host)
2751555560243478560 + 5939149587220868940 = 0 (device)
01111000 10011011 10010110 01001001 01001010 01000101 01101111 01101100
00000000 00000000 00000000 00000000 00000000 00000000 00000000 00000000

$ nvcc -gencode=arch=compute_86,code=sm_86 compressed.cu -o compressed && ./compressed 1024 1024 2
Result 513 of GPU is not the same as CPU result:
507909867887034188 + 5913013224710480936 = 6420923092597515124 (host)
507909867887034188 + 5913013224710480936 = 0 (device)
01011001 00011011 10110011 10001000 10000000 01100000 10100011 01110100
00000000 00000000 00000000 00000000 00000000 00000000 00000000 00000000

$ nvcc -gencode=arch=compute_86,code=sm_86 compressed.cu -o compressed && ./compressed 2048 1024 2
Result 1025 of GPU is not the same as CPU result:
6078473267919993420 + 150973190529220896 = 6229446458449214316 (host)
6078473267919993420 + 150973190529220896 = 0 (device)
01010110 01110011 01110000 10001100 10000111 00101110 00111011 01101100
00000000 00000000 00000000 00000000 00000000 00000000 00000000 00000000

$ nvcc -gencode=arch=compute_86,code=sm_86 compressed.cu -o compressed && ./compressed 8192 512 1
Result of GPU is the same as CPU result.
Result of GPU is the same as CPU result.
Result of GPU is the same as CPU result.
num_elements: 8192, block size: 512, grid_size: 1, runtime: 50.688339ms

$ nvcc -gencode=arch=compute_86,code=sm_86 compressed.cu -o compressed && ./compressed 8192 512 2
Result 4096 of GPU is not the same as CPU result:
2462100123478677000 + 2099582639760097631 = 4561682763238774631 (host)
2462100123478677000 + 2099582639760097631 = 0 (device)
00111111 01001110 01011010 01001101 11101000 11011110 10010111 01100111
00000000 00000000 00000000 00000000 00000000 00000000 00000000 00000000
```

Wrong results at index (one bigger than) of chosen block_size
  in the following calls

```
$ nvcc -gencode=arch=compute_86,code=sm_86 compressed.cu -o compressed && ./compressed 98304 1024 96
Result 1025 of GPU is not the same as CPU result:
4324892774069768041 + 3028766430463943752 = 7353659204533711793 (host)
4324892774069768041 + 3028766430463943752 = 0 (device)
01100110 00001101 01110010 00010101 11001101 10000011 01110011 10110001
00000000 00000000 00000000 00000000 00000000 00000000 00000000 00000000

$ nvcc -gencode=arch=compute_86,code=sm_86 compressed.cu -o compressed && ./compressed 98304 512 192
Result 512 of GPU is not the same as CPU result:
940805045017530629 + 5439509423802620229 = 6380314468820150858 (host)
940805045017530629 + 5439509423802620229 = 0 (device)
01011000 10001011 01101110 00110100 01111101 01001111 01101010 01001010
00000000 00000000 00000000 00000000 00000000 00000000 00000000 00000000
```